agent/requirements.txt
azure-ai-projects>=1.0.0b5
azure-identity>=1.19.0
